bogus fixed types removed from libnds

+++ Graphics/3D/nehe/lesson10/source/main.cpp
@@ -53,14 +53,14 @@ typedef struct tagSECTOR
 
 float sin(float angle)
 {
-	f32 s = SIN[(int)((angle * LUT_SIZE) / 360.0) & LUT_MASK];
+	int32 s = SIN[(int)((angle * LUT_SIZE) / 360.0) & LUT_MASK];
 
 	return f32tofloat(s);
 }
 
 float cos(float angle)
 {
-	f32 c = COS[(int)((angle * LUT_SIZE) / 360.0) & LUT_MASK];
+	int32 c = COS[(int)((angle * LUT_SIZE) / 360.0) & LUT_MASK];
 
 	return f32tofloat(c);
 }
