+++ Graphics/3D/nehe/lesson10b/source/main.cpp
@@ -20,11 +20,11 @@
 int DrawGLScene();
 
 int heading;
-f32 xpos;
-f32 zpos;
+int32 xpos;
+int32 zpos;
 
 int	yrot;				// Y Rotation
-f32 walkbias = 0;
+int32 walkbias = 0;
 int walkbiasangle = 0;
 int lookupdown = 0;
 
@@ -254,9 +254,9 @@ int DrawGLScene()											// Here's Where We Do All The Drawing
 
 	v16 x_m, y_m, z_m;
 	t16 u_m, v_m;
-	f32 xtrans = -xpos;
-	f32 ztrans = -zpos;
-	f32 ytrans = -walkbias-(1<<10);
+	int32 xtrans = -xpos;
+	int32 ztrans = -zpos;
+	int32 ytrans = -walkbias-(1<<10);
 	int sceneroty = LUT_SIZE - yrot;
 	
 	glLoadIdentity();
